RAG_APP/requirements.txt
streamlit
pypdf2
langchain
#langchain-openai
langchain-community
python-dotenv
faiss-cpu
openai
huggingface_hub
InstructorEmbedding
sentence-transformers
